Add ExecutorService for metric retrieval
This service will add parallelism to the retrieval of metrics
from the index_metrics table, decreasing the overall planning time.
A new property, index_short_circuit_cardinality_fetch, is used to enable
a new feature which causes the metrics retrieval to return early if one
of the cardinalities retrieved is less than a given threshold.  The
remaining tasks will complete as planned, caching the results in the
Guava cache, but the connector can use the smallest cardinality
column in the meantime.