Fix an issue when all selected columns of a row are null
The cause of this issue is when the AccumuloRecordCursor scans data,
it sets the scanner to fetch the columns based on the query.  Since
NULL values are not stored in Accumulo (the absence of an entry is
NULL), it is possible that no data will be scanned.  When no data is
returned by Accumulo, we cannot recreate the row ID and the remaining
values of NULL.  By duplicating the row ID in a special entry and
configuring the scanner to retrieve this column (among the others),
we are guaranteed to receive at least one entry and return the correct
data to the client.  This effectively removes the need for at least
two columns to be defined.