Add column visibility to the "number of rows" metric
We're making an assumption for the "number of rows" metric that this mutation has homogenous ColumnUpdate values.  As this is for relational data, in practice, this will generally be true.  This allows for more accurate metrics instead of lumping them all into an entry with no column visibility (not to mention the security issues associated with that).