Add LoadingCache for IndexLookup
This LoadingCache has a short expiration, intended to shorten load times
when concurrent queries contain the same predicates.  Instead of
multiple threads loading the index from Accumulo, only one thread
will load the index while the other threads wait.  The LoadingCache
will expire the entries after 10 seconds -- it is not intended to be
an actual cache, we are just using the blocking behavior in the load
operation.